fix image timeout

+++ app/src/main/java/com/kivoa/controlhub/data/ProductImage.kt
@@ -3,8 +3,8 @@ package com.kivoa.controlhub.data
 import com.squareup.moshi.Json
 
 data class ProductImage(
-    @Json(name = "id") val id: Int,
-    @Json(name = "product_id") val productId: Int,
+    @Json(name = "id") val id: Long,
+    @Json(name = "product_id") val productId: Long,
     @Json(name = "image_url") val imageUrl: String,
     @Json(name = "status") val status: String,
     @Json(name = "created_at") val createdAt: String,
